requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-telegram-bot==20.7
redis==5.0.1
celery==5.3.4
# MongoDB async driver (already included above)
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0

# AI/ML Dependencies - Updated for Python 3.12
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.36.2
peft>=0.7.1
accelerate>=0.25.0
bitsandbytes>=0.41.3
motor>=3.3.2
pymongo>=4.6.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
datasets>=2.14.0
scikit-learn>=1.3.2
pandas>=2.1.4
numpy>=1.25.2
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.17.0
ta>=0.10.2
yfinance>=0.2.28
ccxt>=4.1.77
openai>=1.3.0

# Blockchain
web3>=6.11.3
eth-account>=0.9.0
walletconnect>=0.0.1

# Database
motor>=3.3.2
pymongo>=4.6.0

# Email validation handled by custom validators in models

# Monitoring & Logging
structlog>=23.2.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0

aiohttp==3.9.1
# Advanced Learning Dependencies
talib>=0.4.28
optuna>=3.4.0

# Admin Panel & Monitoring Dependencies
sentry-sdk[fastapi]>=1.38.0
psutil>=5.9.0
